wip(ui): move filterContext outside pages folder
