+++ ui/common/utils/routing.js
@@ -1,4 +1,4 @@
-import { stateToQueryString } from "../../pages/app/visualiser-les-indicateurs/FiltersContext";
+import { stateToQueryString } from "../../components/_pagesComponents/FiltersContext.js";
 import { NAVIGATION_PAGES } from "../constants/navigationPages";
 
 export const navigateToOrganismePage = (history, organisme) => {
