+++ ui/components/TerritoireFilter/TerritoireFilter.jsx
@@ -1,7 +1,7 @@
 import PropTypes from "prop-types";
 import React, { useState } from "react";
 
-import { filtersPropTypes } from "../../pages/app/visualiser-les-indicateurs/FiltersContext";
+import { filtersPropTypes } from "../../components/_pagesComponents/FiltersContext.js";
 import { OverlayMenu, PrimarySelectButton } from "..";
 import SecondarySelectButton from "../SelectButton/SecondarySelectButton";
 import { TERRITOIRE_TYPE } from "./constants";
