+++ ui/components/tables/RepartitionEffectifsParDepartement/CfaRow.jsx
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import React from "react";
 
 import { isDateFuture } from "../../../common/utils/dateUtils";
-import { useFiltersContext } from "../../../pages/app/visualiser-les-indicateurs/FiltersContext";
+import { useFiltersContext } from "../../../components/_pagesComponents/FiltersContext.js";
 import { mapNatureOrganismeDeFormation } from "../../../pages/app/visualiser-les-indicateurs/par-organisme/sections/informations-cfa/CfaInformationSection";
 import NatureOrganismeDeFormationWarning from "../../NatureOrganismeDeFormationWarning/NatureOrganismeDeFormationWarning";
 import NumberValueCell from "../NumberValueCell";
