+++ ui/components/tables/RepartitionEffectifsParFormation/AnneeFormationRows.jsx
@@ -7,7 +7,7 @@ import { QUERY_KEYS } from "../../../common/constants/queryKeys";
 import { mapFiltersToApiFormat } from "../../../common/utils/mapFiltersToApiFormat";
 import { pick } from "../../../common/utils/pick";
 import { sortAlphabeticallyBy } from "../../../common/utils/sortAlphabetically";
-import { useFiltersContext } from "../../../pages/app/visualiser-les-indicateurs/FiltersContext";
+import { useFiltersContext } from "../../../components/_pagesComponents/FiltersContext.js";
 import AnneeFormationRow from "./AnneeFormationRow";
 
 const AnneeFormationRows = ({ formationCfd }) => {
