+++ ui/components/tables/RepartitionEffectifsParFormation/FormationRows.jsx
@@ -7,7 +7,7 @@ import { QUERY_KEYS } from "../../../common/constants/queryKeys";
 import { mapFiltersToApiFormat } from "../../../common/utils/mapFiltersToApiFormat";
 import { pick } from "../../../common/utils/pick";
 import { sortAlphabeticallyBy } from "../../../common/utils/sortAlphabetically";
-import { useFiltersContext } from "../../../pages/app/visualiser-les-indicateurs/FiltersContext";
+import { useFiltersContext } from "../../../components/_pagesComponents/FiltersContext.js";
 import RowsSkeleton from "../../skeletons/RowsSkeleton";
 import FormationRow from "./FormationRow";
