+++ ui/components/tables/RepartitionEffectifsParSiretAndDepartement/EffectifBySiretRow.jsx
@@ -2,7 +2,7 @@ import { Box, Link, Td, Tr } from "@chakra-ui/react";
 import PropTypes from "prop-types";
 import React from "react";
 
-import { useFiltersContext } from "../../../pages/app/visualiser-les-indicateurs/FiltersContext";
+import { useFiltersContext } from "../../../components/_pagesComponents/FiltersContext.js";
 import NumberValueCell from "../NumberValueCell";
 
 const EffectifBySiretRow = ({ siret_etablissement, nom_etablissement, effectifs, isPeriodInvalid }) => {
